applying offset to registers, calibrating z

+++ types/types.go
@@ -83,9 +83,9 @@ func (a *Sensor) SetConfig(config Config) {
 }
 
 type Offsets struct {
-	X float64 `yaml:"X"`
-	Y float64 `yaml:"Y"`
-	Z float64 `yaml:"Z"`
+	X int16 `yaml:"X"`
+	Y int16 `yaml:"Y"`
+	Z int16 `yaml:"Z"`
 }
 
 type ImuReadingQualities struct {
